+ Common parent class for multiple related inputs
